chore: add vitest infrastructure and migrate error_tracking
